requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-widget-tweaks==1.4.5
gcloud==0.18.3
googleapis-common-protos==1.6.0
httplib2==0.13.0
idna==2.8
jws==0.1.3
oauth2client==4.1.2
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodome==3.8.2
Pyrebase4==4.3.0
python-jwt==2.0.1
pytz==2019.1
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
